+++ practico1-WebAPI/Shared/Factura.cs
@@ -9,7 +9,7 @@ namespace Shared
     public class Factura
     {
         public int Id { get; set; }
-        public float TotalComisiones { get; set; } 
+        public decimal TotalComisiones { get; set; } 
         public DateTime FechaInicio { get; set; }
         public DateTime FechaFin { get; set; }
         public int EmpresaId { get; set; }
